Add option for Helm `--include-crds`
I consider that a feature for powerusers.

My use case is not “interactive” use of the kubenix CLI but only as a library for generating k8s JSON. I then use that in ArgoCD using a ArgoCD config management plugin which calls `nix build`.

I think `includeCRDs` should be off by default as users of kubenix which turn it on need to carefully check for updates of all included CRDs when updating the chart.

This is the policy that Helm decided on as well: Helm will only initially install CRDs, never update them.